Renamed to project-files to a more fitting name.
