+++ HexDecBin_Calculator.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30621.155
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HexaCalculator", "HexaCalculator\HexaCalculator.csproj", "{0848B279-2FBA-425D-A3EE-F4AF9D7A4130}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HexDecBin_Calculator", "HexDecBin_Calculator\HexDecBin_Calculator.csproj", "{0848B279-2FBA-425D-A3EE-F4AF9D7A4130}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
