+++ HexDecBin_Calculator/App.config
