+++ HexDecBin_Calculator/Form1.resx
