+++ HexDecBin_Calculator/HexDecBin_Calculator.csproj
@@ -6,8 +6,8 @@
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{0848B279-2FBA-425D-A3EE-F4AF9D7A4130}</ProjectGuid>
     <OutputType>WinExe</OutputType>
-    <RootNamespace>HexaCalculator</RootNamespace>
-    <AssemblyName>HexaCalculator</AssemblyName>
+    <RootNamespace>HexDecBin_Calculator</RootNamespace>
+    <AssemblyName>HexDecBin_Calculator</AssemblyName>
     <TargetFrameworkVersion>v4.7.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
@@ -66,6 +66,7 @@
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
       <DependentUpon>Resources.resx</DependentUpon>
+      <DesignTime>True</DesignTime>
     </Compile>
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
