+++ HexDecBin_Calculator/Properties/AssemblyInfo.cs
