+++ HexDecBin_Calculator/Properties/Resources.resx
