+++ HexDecBin_Calculator/Properties/Settings.settings
